requirements.txt
# Core Autogen packages
annotated-types==0.7.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
pyautogen==0.10.0

# ML
numpy==2.3.3
scipy==1.16.1
scikit-learn==1.7.2
matplotlib==3.10.6
pandas>=2.0.0
xgboost==3.0.5

# Visualization
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
pillow==11.3.0

# PDF processing
PyMuPDF==1.25.4

# Async and concurrency
joblib==1.5.2
threadpoolctl==3.6.0

# Data validation
pydantic==2.11.7
pydantic_core==2.33.2

# JSON and configuration
jsonref==1.1.0
pyparsing==3.2.3

# Utilities
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
typing-inspection==0.4.1
importlib_metadata==8.7.0
zipp==3.23.0

# Monitoring and observability
opentelemetry-api==1.36.0
protobuf==5.29.5

# Additional packages
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.8.0
tiktoken>=0.5.0
chromadb>=0.4.0
langchain>=0.1.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0